+++ AddOns/MinimalistCooldownEdge/Modules/Styler.lua
@@ -10,7 +10,7 @@ local MCE = LibStub("AceAddon-3.0"):GetAddon(C.Addon.AceName)
 local Styler = MCE:NewModule("Styler")
 
 local wipe = wipe
-local C_Timer_After = C_Timer.After
+local RunAfter = addon.RunAfter
 
 
 local Registry, BatchProcessor, StyleEngine, DurationColor, Classifier
@@ -32,7 +32,7 @@ function Styler:OnEnable()
     end)
 
     -- Initial adapter-driven discovery (short delay so frames exist)
-    C_Timer_After(2, function()
+    RunAfter(2, function()
         self:ForceUpdateAll(true)
     end)
 end